move useCurrencyBalance to app/hooks
