+++ apps/flame-defi/app/hooks/use-currency-balance.ts
